insercao de TEC que havia sido esquecido

+++ src/visao/FormDuracaoChamada.java
@@ -23,7 +23,7 @@ public class FormDuracaoChamada extends javax.swing.JFrame {
      * Creates new form FormDuracaoChamada
      */
     private Distribuicao distrib;
-    private boolean definido;
+        private boolean definido;
     
     public FormDuracaoChamada() {
         initComponents();  
@@ -209,35 +209,35 @@ public class FormDuracaoChamada extends javax.swing.JFrame {
 
     private void jButton1MouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jButton1MouseClicked
         int opt = jComboBox1.getSelectedIndex();
-        int a,b,c = 0;
+        double a,b,c = 0;
         try{            
             switch(opt){
                 case 0:
-                    a = Integer.parseInt(jTextField1.getText());
-                    b = Integer.parseInt(jTextField2.getText());
-                    c = Integer.parseInt(jTextField3.getText());
+                    a = Double.parseDouble(jTextField1.getText());
+                    b = Double.parseDouble(jTextField2.getText());
+                    c = Double.parseDouble(jTextField3.getText());
                     distrib = new Triangular(a, c, b);
                     definido = true;
                     break;
                 case 1:
-                    a = Integer.parseInt(jTextField1.getText());
-                    b = Integer.parseInt(jTextField2.getText());
+                    a = Double.parseDouble(jTextField1.getText());
+                    b = Double.parseDouble(jTextField2.getText());
                     distrib = new Normal(a,b);
                     definido = true;
                     break;
                 case 2:
-                    a = Integer.parseInt(jTextField1.getText());
+                    a = Double.parseDouble(jTextField1.getText());
                     distrib = new Exponencial(a);
                     definido = true;
                     break;
                 case 3:
-                    a = Integer.parseInt(jTextField1.getText());
-                    b = Integer.parseInt(jTextField2.getText());
+                    a = Double.parseDouble(jTextField1.getText());
+                    b = Double.parseDouble(jTextField2.getText());
                     distrib = new Uniforme(a,b);
                     definido = true;
                     break;
                 case 4:
-                    a = Integer.parseInt(jTextField1.getText());
+                    a = Double.parseDouble(jTextField1.getText());
                     distrib = new Constante(a);
                     definido = true;
                     break;
